the custom JisonLexerError and JisonParserError error exception classes have been migrated to ES6 code. An updated version of the ES5 equivalent is still included (in comments) in the code for use by anyone who needs it, but our general direction is to phase out ES5 (after all, that's what Babel is for, when you need to support antiquated browsers, etc.)
The relevant tests in the lexer have been tweaked a bit, where the only tests that needed (minimal) change are those testing this type of Error with a *nested* Error instance -- a scenario which can happen in the generated parser when some severe error in the userland action code (or similar) causes a generic exception / Error to be trown from within the running parser. Note the *parser kernel code* where such a scenario is checked for near the end, in the `catch` clause, where the relevant code ensures that `parseError()` is only invoked *once* per thrown error/exception.